Chore: Move categories queries to inside the category component
