+++ react/components/CategoryTranslation/graphql/getAllCategories.gql
