+++ react/components/CategoryTranslation/graphql/getCategory.gql
