+++ react/components/CategoryTranslation/graphql/translateCategory.gql
