+++ react/components/CategoryTranslation/graphql/uploadCategoryTranslations.gql
